requirements.txt
chainlit
python-dotenv
openai-agents
websockets
google-generativeai
